Update tests, change is_streamable
