+++ include/yaml-cpp/exceptions.h
@@ -123,7 +123,7 @@ inline const std::string BAD_SUBSCRIPT_WITH_KEY(
 
 inline const std::string BAD_SUBSCRIPT_WITH_KEY(const std::string& key) {
   std::stringstream stream;
-  stream << BAD_SUBSCRIPT << " (key: '" << key << "')";
+  stream << BAD_SUBSCRIPT << " (key: \"" << key << "\")";
   return stream.str();
 }
 
@@ -131,7 +131,7 @@ template <typename T>
 inline const std::string BAD_SUBSCRIPT_WITH_KEY(
     const T& key, typename enable_if<is_numeric<T>>::type* = 0) {
   std::stringstream stream;
-  stream << BAD_SUBSCRIPT << " (key: '" << key << "')";
+  stream << BAD_SUBSCRIPT << " (key: \"" << key << "\")";
   return stream.str();
 }
 
@@ -140,7 +140,7 @@ inline const std::string INVALID_NODE_WITH_KEY(const std::string& key) {
   if (key.empty()) {
     return INVALID_NODE;
   }
-  stream << "invalid node - first invalid key: '" << key << "'";
+  stream << "invalid node; first invalid key: \"" << key << "\"";
   return stream.str();
 }
 }
