+++ include/yaml-cpp/node/node.h
@@ -132,7 +132,7 @@ class YAML_CPP_API Node {
 
  private:
   bool m_isValid;
-  // string representation of invalid key, if the node is invalid
+  // String representation of invalid key, if the node is invalid.
   std::string m_invalidKey;
   mutable detail::shared_memory_holder m_pMemory;
   mutable detail::node* m_pNode;
