+++ test/integration/error_messages_test.cpp
@@ -1,4 +1,3 @@
-#include "specexamples.h"
 #include "yaml-cpp/yaml.h"  // IWYU pragma: keep
 
 #include "gtest/gtest.h"
@@ -14,15 +13,19 @@
 namespace YAML {
 namespace {
 
-TEST(ErrorMessageTest, Ex9_1_BadSubscriptErrorMessage) {
-  Node doc = Load(ex9_1);
+TEST(ErrorMessageTest, BadSubscriptErrorMessage) {
+  const char *example_yaml = "first:\n"
+                             "   second: 1\n"
+                             "   third: 2\n";
+
+  Node doc = Load(example_yaml);
 
   // Test that printable key is part of error message
   EXPECT_THROW_EXCEPTION(YAML::BadSubscript, doc["first"]["second"]["fourth"],
-                         "operator[] call on a scalar (key: 'fourth')");
+                         "operator[] call on a scalar (key: \"fourth\")");
   
   EXPECT_THROW_EXCEPTION(YAML::BadSubscript, doc["first"]["second"][37],
-                         "operator[] call on a scalar (key: '37')");
+                         "operator[] call on a scalar (key: \"37\")");
 
 
   // Non-printable key is not included in error message
@@ -35,14 +38,18 @@ TEST(ErrorMessageTest, Ex9_1_BadSubscriptErrorMessage) {
 }
 
 TEST(ErrorMessageTest, Ex9_1_InvalidNodeErrorMessage) {
-  const Node doc = Load(ex9_1);
+  const char *example_yaml = "first:\n"
+                             "   second: 1\n"
+                             "   third: 2\n";
+
+  const Node doc = Load(example_yaml);
 
   // Test that printable key is part of error message
   EXPECT_THROW_EXCEPTION(YAML::InvalidNode, doc["first"]["fourth"].as<int>(),
-                         "invalid node - first invalid key: 'fourth'");
+                         "invalid node; first invalid key: \"fourth\"");
   
   EXPECT_THROW_EXCEPTION(YAML::InvalidNode, doc["first"][37].as<int>(),
-                         "invalid node - first invalid key: '37'");
+                         "invalid node; first invalid key: \"37\"");
  
   // Non-printable key is not included in error message
   EXPECT_THROW_EXCEPTION(YAML::InvalidNode,
